# backend/src/config/mobileOpenApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lidifin Mobile API",
    "version": "1.0.0",
    "description": "Dedicated mobile-facing API contract for native clients. All paths are relative to /api. Primary authentication is username/password via JWT access + refresh tokens."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Primary mobile authentication. Obtain with POST /auth/login and refresh with POST /auth/refresh."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Secondary device authentication for paired devices and integrations."
      },
      "streamTokenAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Optional query token for stream URLs when attaching Authorization headers is inconvenient."
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Web UI session authentication. Included for completeness."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": {}
        }
      },
...
      "ApiKey": {
        "type": "object",
        "required": ["id", "name", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastUsed": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api-keys": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Create a new API key",
        "description": "Creates a long-lived API key for paired devices or integrations. This is not the primary mobile sign-in flow.",
        "security": [
          { "bearerAuth": [] },
          { "sessionAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["deviceName"],
                "properties": {
                  "deviceName": { "type": "string", "example": "Pixel 9" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": { "type": "string" },
                    "name": { "type": "string" },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["API Keys"],
        "summary": "List API keys",
        "security": [
          { "bearerAuth": [] },
          { "sessionAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "API keys returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKeys": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ApiKey" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "tags": ["API Keys"],
        "summary": "Revoke an API key",
        "security": [
          { "bearerAuth": [] },
          { "sessionAuth": [] }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}